--- streamlit ui/.dockerignore
@@ -1,2 +1,0 @@
-# Ignore the .streamlit/secrets.toml file
-.streamlit/secrets.toml
