+++ streamlit_ui/.dockerignore
@@ -1,0 +1,2 @@
+# Ignore the .streamlit/secrets.toml file
+.streamlit/secrets.toml
